requirements.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.5
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.1
httpie==1.0.2
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mysqlclient==1.4.1
openapi-codec==1.3.2
Pygments==2.3.1
pytz==2018.9
requests==2.21.0
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
